requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pyyaml==6.0.1
tldextract==5.1.1
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2
